task 6

+++ 0x00-hello_world/6-size.c
@@ -8,7 +8,7 @@ int main(void)
 char c;
 int i;
 long l;
-long int li;
+long long int li;
 float fl;
 printf("Size of a char:%lu byte(s)\n", sizeof(c));
 printf("Size of an int:%lu byte(s)\n", sizeof(i));
